mm-ai-agent/requirements.txt
# Web framework
fastapi==0.105.0
uvicorn==0.24.0.post1
pydantic==2.5.2
starlette==0.27.0

# HTTP client
aiohttp==3.9.1
httpx==0.25.2

# Environment & config
python-dotenv==1.0.0
pyyaml==6.0.1

# LLM libraries
langchain==0.0.340
langchain-core==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.1
langchain-google-genai==0.0.5
openai==1.3.8

# Caching & storage
redis==5.0.1
aioredis==2.0.1

# Security
pyjwt==2.8.0
bcrypt==4.1.1
cryptography==41.0.5

# Utilities
python-multipart==0.0.6
tenacity==8.2.3
aiodns==3.1.1
cchardet==2.1.7

# Monitoring & logging
prometheus-client==0.18.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0